+++ app/Http/Controllers/Admin/TernakController.php
@@ -232,7 +232,13 @@ class TernakController extends Controller
     public function destroy($id)
     {
         $data = Ternak::findOrFail($id);
-        $data->delete();
+        if(Perkawinan::where('necktag', $id)->exists()){
+            $err = 'Data ternak id '. $id .' tidak dapat dihapus.';
+            return response()->json(['error' => $err]);
+        }
+        else{
+            $data->delete();
+        }
     }
